Move IP address selection to its own file
